requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
Werkzeug==2.3.7
psycopg2-binary==2.9.10
stripe==7.5.0
flasgger==0.9.7.1